Base geometry on screen instead of workspace
Using the workspace area is great for some things, but since this window
isn't actually being drawn IN the workspace it doesn't make much sense.
Worse, since one screen has a Wibox and the other doesn't, the quake
panels were rendering a different size on each of my identical screens.